#75 Simplify badge views by introducing AttendeeBadge view
