+++ DevoxxClientCommon/src/main/resources/com/devoxx/devoxx.properties
@@ -36,9 +36,9 @@ OTN.VIEW.NOTES=My Notes
 OTN.VIEW.BADGE=Badge Details
 OTN.VIEW.BADGES=My Badges
 OTN.VIEW.VENUE=Venue
-OTN.VIEW.SPONSOR=Sponsor Login
-OTN.VIEW.ATTENDEE=Attendee Login
 OTN.VIEW.SPONSORS=Select Sponsor
+OTN.VIEW.SPONSORBADGE=Sponsor Login
+OTN.VIEW.ATTENDEEBADGE=Attendee Login
 OTN.VIEW.EXHIBITIONMAPS=Exhibition Maps
 OTN.VIEW.EXHIBITIONMAP=Exhibitor Map
 OTN.VIEW.EXHIBITOR=Exhibitor
