+++ DevoxxClientCommon/src/main/resources/com/devoxx/devoxx_fr.properties
@@ -36,8 +36,10 @@ OTN.VIEW.NOTES=Mes notes
 OTN.VIEW.BADGE=D\u00e9tails du badge
 OTN.VIEW.BADGES=Mes Badges
 OTN.VIEW.VENUE=Lieu
-OTN.VIEW.SPONSOR=Sponsor Login
 OTN.VIEW.SPONSORS=S\u00e9lection du sponsor
+#TODO: Translate next 2 to French
+OTN.VIEW.SPONSORBADGE=Sponsor Login
+OTN.VIEW.ATTENDEEBADGE=Attendee Login
 OTN.VIEW.EXHIBITIONMAPS=Plan
 OTN.VIEW.EXHIBITIONMAP=Plan des Exposants
 OTN.VIEW.EXHIBITOR=Exposant
