+++ DevoxxClientMobile/src/main/java/com/devoxx/DevoxxView.java
@@ -48,7 +48,8 @@ public class DevoxxView {
     public static final AppView EXHIBITOR       = view( ExhibitorPresenter.class,      MaterialDesignIcon.PLACE);
     public static final AppView EXHIBITION_MAPS = view( ExhibitionMapsPresenter.class, MaterialDesignIcon.MAP,                SHOW_IN_DRAWER);
     public static final AppView EXHIBITION_MAP  = view( ExhibitionMapPresenter.class,  MaterialDesignIcon.MAP);
-    public static final AppView SPONSORS        = view( SponsorsPresenter.class,       MaterialDesignIcon.CARD_MEMBERSHIP,   SKIP_VIEW_STACK);
+    public static final AppView SPONSORS        = view( SponsorsPresenter.class,       MaterialDesignIcon.CARD_MEMBERSHIP);
+    public static final AppView ATTENDEE_BADGE  = view( AttendeeBadgePresenter.class,  MaterialDesignIcon.CARD_MEMBERSHIP);
     public static final AppView SPONSOR_BADGE   = view( SponsorBadgePresenter.class,   MaterialDesignIcon.CARD_MEMBERSHIP);
     public static final AppView VENUE           = view( VenuePresenter.class,          MaterialDesignIcon.ACCESSIBILITY,      SHOW_IN_DRAWER);
     public static final AppView NOTES           = view( NotesPresenter.class,          MaterialDesignIcon.EVENT_NOTE,         SHOW_IN_DRAWER);
