+++ DevoxxClientMobile/src/main/resources/com/devoxx/views/attendeebadge.fxml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import com.gluonhq.charm.glisten.mvc.View?>
+<View fx:id="attendeeView" styleClass="attendee-view" xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml" fx:controller="com.devoxx.views.AttendeeBadgePresenter">
+</View>
